Add files via upload
GroceryPlanner.java is my main file, Item.java is the item class that has the getters and setters for item type, item isle, and special location. There are text files that I am reading in this code that are a direct copy and paste from HEB websites that indicate where all items are depending on store location.